mutt_file_tidy_path: move function to library
