+++ mutt/file.c
@@ -1388,3 +1388,137 @@ int mutt_file_check_empty(const char *path)
 
   return ((st.st_size == 0));
 }
+
+/**
+ * mutt_file_tidy_path - Tidy a path with optional symlink resolution
+ * @param buf Buffer containing path
+ * @param rsym Boolean specifying whether to resolve symlinks
+ * @retval len String length of resolved path
+ * @retval 0   Error, buf is not overwritten
+ *
+ * Resolve and overwrite the path in buf.
+ *
+ * @note Size of buf should be at least PATH_MAX bytes.
+ */
+size_t mutt_file_tidy_path(char *buf, bool rsym)
+{
+  char s[PATH_MAX];
+
+  if (rsym)
+  {
+    char *rp = realpath(buf, s);
+    if (rp)
+      return mutt_str_strfcpy(buf, rp, PATH_MAX);
+    else
+      return mutt_str_strlen(buf);
+  }
+  else
+  {
+    size_t len = 0;
+
+    /* cleanup path - the below "if" migrated from mutt_pretty_mailbox() */
+    if (strstr(buf, "//") || strstr(buf, "/./"))
+    {
+      /* first attempt to collapse the pathname, this is more
+       * lightweight than realpath() and doesn't resolve links
+       */
+      char *q = buf, *r = buf;
+      while (*r)
+      {
+        if ((r[0] == '/') && (r[1] == '/'))
+        {
+          *q++ = '/';
+          r += 2;
+        }
+        else if ((r[0] == '/') && (r[1] == '.') && (r[2] == '/'))
+        {
+          *q++ = '/';
+          r += 3;
+        }
+        else
+          *q++ = *r++;
+      }
+      *q = 0;
+    }
+
+    /* Fix any combo of parent paths. This works by tracking directory
+     * portions via pointers. These pointers are updated as parent paths
+     * are encountered.
+     */
+    if (strstr(buf, ".."))
+    {
+      // How many levels of directories to track
+      const size_t MUTT_PATH_MAX_DEPTH = 100;
+      // The directory level tracker (array of pointers)
+      /* This tracks which substrings of the path belong to which depth in
+       * the directory tree.
+       */
+      char *level[MUTT_PATH_MAX_DEPTH];
+      /* The following loop works such that if a parent directory is encountered,
+       * then the current level n is decremented so that the next path substring
+       * will replace the previous one in the level array. Otherwise, the
+       * level is incremented.
+       *
+       * Given the current nature of the loop, if a parent directory is
+       * encountered last, the level n will be one lower than the actual
+       * "current" level. Otherwise, the level will be one higher than "current".
+       * So, some adjusting is necessary after the loop exits. This could
+       * possibly be optimized out with a different loop design.
+       */
+      bool parent_last = false;
+      char *r = buf; // iterator
+      size_t n = 0;  // current level
+
+      while ((r = strchr(r, '/')) && (r[1] != '\0'))
+      {
+        if ((r[1] == '.') && (r[2] == '.') && ((r[3] == '/') || (r[3] == '\0')))
+        {
+          // decrease dir level -- next path will go to this level
+          if (n > 0)
+            n--;
+          parent_last = true;
+        }
+        else
+        {
+          level[n] = r;
+          if (n < MUTT_PATH_MAX_DEPTH)
+            n++;
+          parent_last = false;
+        }
+        r++;
+      }
+
+      r = buf; // iterator from start of path
+
+      if (parent_last && (n == 0)) // The parent of root, is simply root
+      {
+        r[0] = '/';
+        r[1] = '\0';
+      }
+      else
+      {
+        for (int i = 0; i < n; i++) // iterate through tracked dir hierarchy
+        {
+          *(r++) = level[i][0]; // write the first character ('/')
+          // write the path component until '/' or end-of-string
+          for (char *j = &level[i][1]; (*j != '/') && (*j != '\0'); j++)
+            *(r++) = *j;
+        }
+        r[0] = '\0';
+      }
+    }
+
+    /* Let's cut off the trailing / if it exists and we're not at the root
+     * Note: doing this because the browser GUI is a lil' buggy and if it
+     * sees a path '/a/b/' going to the parent only changes it to '/a/b',
+     * causing the user to have press '..' twice. Honestly, this should be
+     * fixed in the browser but I don't mind removing trailing slashes
+     * from all paths that come through this (non-symlinking) code-path.
+     */
+    len = strlen(buf);
+    if ((len > 1) && (buf[len - 1] == '/'))
+      buf[--len] = '\0';
+
+    return len;
+  }
+}
